feat(imageloader): expose generic setting struct
