+++ include/tev/imageio/IcoImageLoader.h
@@ -28,7 +28,7 @@ namespace tev {
 class IcoImageLoader : public ImageLoader {
 public:
     Task<std::vector<ImageData>> load(
-        std::istream& iStream, const fs::path& path, std::string_view channelSelector, int priority, const GainmapHeadroom& gainmapHeadroom
+        std::istream& iStream, const fs::path& path, std::string_view channelSelector, const ImageLoaderSettings& settings, int priority
     ) const override;
 
     std::string name() const override { return "ICO"; }
