+++ src/Image.cpp
@@ -1408,7 +1408,7 @@ Task<nanogui::Vector2i>
 }
 
 Task<vector<shared_ptr<Image>>> tryLoadImage(
-    int taskPriority, fs::path path, istream& iStream, string_view channelSelector, const GainmapHeadroom& gainmapHeadroom, bool groupChannels
+    int taskPriority, fs::path path, istream& iStream, string_view channelSelector, const ImageLoaderSettings& settings, bool groupChannels
 ) {
     const auto handleException = [&](const exception& e) {
         if (channelSelector.empty()) {
@@ -1445,7 +1445,7 @@ Task<vector<shared_ptr<Image>>> tryLoadImage(
         for (const auto& imageLoader : ImageLoader::getLoaders()) {
             try {
                 loadMethod = imageLoader->name();
-                imageData = co_await imageLoader->load(iStream, path, channelSelector, taskPriority, gainmapHeadroom);
+                imageData = co_await imageLoader->load(iStream, path, channelSelector, settings, taskPriority);
                 success = true;
                 break;
             } catch (const ImageLoader::FormatNotSupported& e) {
@@ -1508,12 +1508,12 @@ Task<vector<shared_ptr<Image>>> tryLoadImage(
 }
 
 Task<vector<shared_ptr<Image>>>
-    tryLoadImage(fs::path path, istream& iStream, string_view channelSelector, const GainmapHeadroom& gainmapHeadroom, bool groupChannels) {
-    co_return co_await tryLoadImage(-Image::drawId(), path, iStream, channelSelector, gainmapHeadroom, groupChannels);
+    tryLoadImage(fs::path path, istream& iStream, string_view channelSelector, const ImageLoaderSettings& settings, bool groupChannels) {
+    co_return co_await tryLoadImage(-Image::drawId(), path, iStream, channelSelector, settings, groupChannels);
 }
 
 Task<vector<shared_ptr<Image>>>
-    tryLoadImage(int taskPriority, fs::path path, string_view channelSelector, const GainmapHeadroom& gainmapHeadroom, bool groupChannels) {
+    tryLoadImage(int taskPriority, fs::path path, string_view channelSelector, const ImageLoaderSettings& settings, bool groupChannels) {
     try {
         path = fs::absolute(path);
     } catch (const runtime_error&) {
@@ -1522,128 +1522,12 @@ Task<vector<shared_ptr<Image>>>
     }
 
     ifstream fileStream{path, ios_base::binary};
-    co_return co_await tryLoadImage(taskPriority, path, fileStream, channelSelector, gainmapHeadroom, groupChannels);
+    co_return co_await tryLoadImage(taskPriority, path, fileStream, channelSelector, settings, groupChannels);
 }
 
 Task<vector<shared_ptr<Image>>>
-    tryLoadImage(fs::path path, string_view channelSelector, const GainmapHeadroom& gainmapHeadroom, bool groupChannels) {
-    co_return co_await tryLoadImage(-Image::drawId(), path, channelSelector, gainmapHeadroom, groupChannels);
-}
-
-void BackgroundImagesLoader::enqueue(const fs::path& path, string_view channelSelector, bool shallSelect, const shared_ptr<Image>& toReplace) {
-    // If we're trying to open a directory, try loading all the images inside of that directory
-    if (fs::exists(path) && fs::is_directory(path)) {
-        tlog::info() << fmt::format("Loading images {}from directory {}", mRecursiveDirectories ? "recursively " : "", path);
-
-        const fs::path canonicalPath = fs::canonical(path);
-        mDirectories[canonicalPath].emplace(channelSelector);
-
-        vector<fs::directory_entry> entries;
-        forEachFileInDir(mRecursiveDirectories, canonicalPath, [&](const auto& entry) {
-            if (!entry.is_directory()) {
-                mFilesFoundInDirectories.emplace(PathAndChannelSelector{entry, string{channelSelector}});
-                entries.emplace_back(entry);
-            }
-        });
-
-        // Open directory entries in natural order (e.g. "file1.exr", "file2.exr", "file10.exr" instead of "file1.exr", "file10.exr"),
-        // selecting the first one.
-        sort(begin(entries), end(entries), [](const auto& a, const auto& b) { return naturalCompare(a.path().string(), b.path().string()); });
-
-        for (size_t i = 0; i < entries.size(); ++i) {
-            enqueue(entries[i], channelSelector, i == 0 ? shallSelect : false);
-        }
-
-        return;
-    }
-
-    // We want to measure the time it takes to load a whole batch of images. Start measuring when the loader queue goes from empty to
-    // non-empty and stop measuring when the queue goes from non-empty to empty again.
-    if (mUnsortedLoadCounter == mLoadCounter) {
-        mLoadStartTime = chrono::system_clock::now();
-        mLoadStartCounter = mUnsortedLoadCounter;
-    }
-
-    const int loadId = mUnsortedLoadCounter++;
-    invokeTaskDetached([loadId, path, channelSelector = string{channelSelector}, shallSelect, toReplace, this]() -> Task<void> {
-        const int taskPriority = -Image::drawId();
-
-        co_await ThreadPool::global().enqueueCoroutine(taskPriority);
-        const auto images = co_await tryLoadImage(taskPriority, path, channelSelector, mGainmapHeadroom, mGroupChannels);
-
-        {
-            const lock_guard lock{mPendingLoadedImagesMutex};
-            mPendingLoadedImages.push({loadId, shallSelect, images, toReplace});
-        }
-
-        if (publishSortedLoads()) {
-            redrawWindow();
-        }
-    });
-}
-
-void BackgroundImagesLoader::checkDirectoriesForNewFilesAndLoadThose() {
-    for (const auto& dir : mDirectories) {
-        forEachFileInDir(mRecursiveDirectories, dir.first, [&](const auto& entry) {
-            if (!entry.is_directory()) {
-                for (const auto& channelSelector : dir.second) {
-                    const PathAndChannelSelector p = {entry, channelSelector};
-                    if (!mFilesFoundInDirectories.contains(p)) {
-                        mFilesFoundInDirectories.emplace(p);
-                        enqueue(entry, channelSelector, false);
-                    }
-                }
-            }
-        });
-    }
-}
-
-optional<ImageAddition> BackgroundImagesLoader::tryPop() {
-    const lock_guard lock{mPendingLoadedImagesMutex};
-    return mLoadedImages.tryPop();
-}
-
-optional<nanogui::Vector2i> BackgroundImagesLoader::firstImageSize() const {
-    const lock_guard lock{mPendingLoadedImagesMutex};
-    if (mLoadedImages.empty()) {
-        return nullopt;
-    }
-
-    const ImageAddition& firstImage = mLoadedImages.front();
-    if (firstImage.images.empty()) {
-        return nullopt;
-    }
-
-    return firstImage.images.front()->size();
-}
-
-bool BackgroundImagesLoader::publishSortedLoads() {
-    const lock_guard lock{mPendingLoadedImagesMutex};
-    bool pushed = false;
-    while (!mPendingLoadedImages.empty() && mPendingLoadedImages.top().loadId == mLoadCounter) {
-        // null image pointers indicate failed loads. These shouldn't be pushed.
-        if (!mPendingLoadedImages.top().images.empty()) {
-            mLoadedImages.push(mPendingLoadedImages.top());
-        }
-
-        mPendingLoadedImages.pop();
-        pushed = true;
-
-        ++mLoadCounter;
-    }
-
-    if (mLoadCounter == mUnsortedLoadCounter && mLoadCounter - mLoadStartCounter > 1) {
-        const auto end = chrono::system_clock::now();
-        const chrono::duration<double> elapsedSeconds = end - mLoadStartTime;
-        tlog::success() << fmt::format("Loaded {} images in {:.3f} seconds.", mLoadCounter - mLoadStartCounter, elapsedSeconds.count());
-    }
-
-    return pushed;
-}
-
-bool BackgroundImagesLoader::hasPendingLoads() const {
-    const lock_guard lock{mPendingLoadedImagesMutex};
-    return mLoadCounter != mUnsortedLoadCounter;
+    tryLoadImage(fs::path path, string_view channelSelector, const ImageLoaderSettings& settings, bool groupChannels) {
+    co_return co_await tryLoadImage(-Image::drawId(), path, channelSelector, settings, groupChannels);
 }
 
 } // namespace tev
