+++ src/ImageViewer.cpp
@@ -2474,7 +2474,7 @@ void ImageViewer::pasteImagesFromClipboard() {
 
     tlog::info() << "Loading image from clipboard...";
     auto imagesLoadTask = tryLoadImage(
-        fmt::format("clipboard ({})", ++mClipboardIndex), imageStream, "", mImagesLoader->gainmapHeadroom(), mImagesLoader->groupChannels()
+        fmt::format("clipboard ({})", ++mClipboardIndex), imageStream, "", mImagesLoader->imageLoaderSettings(), mImagesLoader->groupChannels()
     );
 
     const auto images = imagesLoadTask.get();
