+++ src/imageio/ClipboardImageLoader.cpp
@@ -27,7 +27,7 @@ using namespace std;
 
 namespace tev {
 
-Task<vector<ImageData>> ClipboardImageLoader::load(istream& iStream, const fs::path&, string_view, int priority, const GainmapHeadroom&) const {
+Task<vector<ImageData>> ClipboardImageLoader::load(istream& iStream, const fs::path&, string_view, const ImageLoaderSettings&, int priority) const {
     char magic[4];
     clip::image_spec spec;
